Add a wrapper around a socket connection that includes useful helper methods
send guarantees that any data passed to it will be sent, if it is still possible. recieve guarantees that it will read any available data on the socket and chunk it into discrete packets that are added to the list of recieved messages. get_next_request will return the oldest message, if any are ready, or None.