Add an enum of request types, a abstract class NetworkObject to act as a template for other requests, and GetGames and GameList requests
GetGames contains no special data, and is simply a prompt for the server to return a list of running games. GameList is the server's response and contains a list of GameDescriptors for each of its running games. A request can be generated from data or a binary representation and can be converted to a binary representation. The binary representation of a request is always a bytearray with two bytes for length, one for type and the remainer is the special data for that request.