add JoinGame, PlayerMove, and ClientUpdate messages
JoinGame allows a player to request a game lobby to add it to a game after recieving a gamelist. PlayerMove allows a player to inform a game instance of the user's movement inputs. ClientUpdate allows a game instance to inform a player of game updates like other enemies and attacks